+++ flask_restful_dbbase/resources/model_resource.py
@@ -3,11 +3,8 @@
 This module implements a starting point for model resources.
 
 """
-from flask_restful import request
+from flask_restful import request, current_app
 from .dbbase_resource import DBBaseResource
-import logging
-
-logger = logging.getLogger(__file__)
 
 
 class ModelResource(DBBaseResource):
@@ -80,6 +77,7 @@ class ModelResource(DBBaseResource):
                 else:
                     func = self.process_get_input.__name__
                     msg = f"malformed error in {func}: {result}"
+                    current_app.logger.error(msg)
                     return {"message": msg}, 500
 
             query, data = result
@@ -92,7 +90,7 @@ class ModelResource(DBBaseResource):
         except Exception as err:
             msg = err.args[0]
             return_msg = f"Internal Server Error: method {FUNC_NAME}: {url}"
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return {"message": return_msg}, 500
 
         sfields, sfield_relations = self._get_serializations(FUNC_NAME)
@@ -102,12 +100,12 @@ class ModelResource(DBBaseResource):
                 serial_fields=sfields, serial_field_relations=sfield_relations,
             )
 
-            logger.debug(result)
+            current_app.logger.debug(result)
 
             return result, 200
 
         msg = f"{self.model_name} with {kdict} not found"
-        logger.debug(msg)
+        current_app.logger.debug(msg)
         return {"message": msg}, 404
 
     def post(self):
@@ -125,9 +123,11 @@ class ModelResource(DBBaseResource):
             except Exception as err:
                 msg = err
                 return_msg = f"A JSON format problem:{msg}: {request.data}"
+                current_app.logger.error(msg)
                 return {"message": return_msg}, 400
 
         else:
+            current_app.logger.info(msg)
             return {"message": "JSON format is required"}, 415
 
         if self.process_post_input is not None:
@@ -141,6 +141,7 @@ class ModelResource(DBBaseResource):
                 else:
                     func = self.process_post_input.__name__
                     msg = f"malformed error in {func}: {result}"
+                    current_app.logger.error(msg)
                     return {"message": msg}, 500
 
             data = result
@@ -151,6 +152,7 @@ class ModelResource(DBBaseResource):
         )
 
         if status is False:
+            current_app.logger.info(msg)
             return {"message": data}, 400
 
         key_names = self.get_key_names(formatted=False)
@@ -172,12 +174,14 @@ class ModelResource(DBBaseResource):
                 msg = err.args[0]
                 return_msg = f"Internal Server Error: method {FUNC_NAME}: "
                 f"{url}"
-                logger.error(f"{url} method {FUNC_NAME}: {msg}")
+                current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
                 return {"message": return_msg}, 500
 
         if item:
+            msg = f"{kdict} for {self.model_name} already exists."
+            current_app.logger.info(msg)
             return (
-                {"message": f"{kdict} for {self.model_name} already exists."},
+                {"message": msg},
                 409,
             )
 
@@ -235,6 +239,8 @@ class ModelResource(DBBaseResource):
                         skip_missing_data=True,
                     )
                     if sub_status is False:
+                        # NOTE: look at this further
+                        current_app.logger.info(sub_data)
                         return {"message": sub_data}, 400
 
                     getattr(item, key).append(sub_class(**sub_data))
@@ -256,7 +262,7 @@ class ModelResource(DBBaseResource):
         except Exception as err:
             msg = err.args[0]
             return_msg = f"Internal Server Error: method {FUNC_NAME}: {url}"
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return {"message": return_msg}, 500
 
         adjust_after = self.after_commit.get(FUNC_NAME)
@@ -292,6 +298,7 @@ class ModelResource(DBBaseResource):
             kdict = self._check_key(kwargs)
         except Exception as err:
             msg = err.args[0]
+            current_app.logger.info(msg)
             return {"message": msg}, 400
 
         if request.is_json:
@@ -300,6 +307,7 @@ class ModelResource(DBBaseResource):
             except Exception as err:
                 msg = err
                 return_msg = f"A JSON format problem:{msg}: {request.data}"
+                current_app.logger.info(msg)
                 return {"message": return_msg}, 400
         else:
             return {"message": "JSON format is required"}, 415
@@ -316,6 +324,7 @@ class ModelResource(DBBaseResource):
                 else:
                     func = self.process_put_input.__name__
                     msg = f"malformed error in {func}: {result}"
+                    current_app.logger.info(msg)
                     return {"message": msg}, 500
 
             query, data = result
@@ -332,7 +341,7 @@ class ModelResource(DBBaseResource):
         except Exception as err:
             msg = err.args[0]
             return_msg = f"Internal Server Error: method {FUNC_NAME}: {url}"
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return {"message": return_msg}, 500
 
         data = self.model_class.deserialize(data)
@@ -343,6 +352,7 @@ class ModelResource(DBBaseResource):
             self.model_class.deserialize(data), self.get_obj_params()
         )
         if status is False:
+            current_app.logger.info(f"{msg}: 400")
             return {"message": data}, 400
 
         if item is None:
@@ -368,7 +378,7 @@ class ModelResource(DBBaseResource):
             self.model_class.db.session.rollback()
             msg = err.args[0]
             return_msg = f"Internal Server Error: method {FUNC_NAME}: {url}"
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return (
                 {
                     "message": "An error occurred updating the "
@@ -446,7 +456,7 @@ class ModelResource(DBBaseResource):
         except Exception as err:
             msg = err.args[0]
             return_msg = f"Internal Server Error: method {FUNC_NAME}: {url}"
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return {"message": return_msg}, 500
 
         data = self.model_class.deserialize(data)
@@ -480,7 +490,7 @@ class ModelResource(DBBaseResource):
         except Exception as err:
             msg = err.args[0]
             self.model_class.db.session.rollback()
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return (
                 {
                     "message": "An error occurred updating the "
@@ -549,12 +559,12 @@ class ModelResource(DBBaseResource):
                 f"Internal Server Error: method {FUNC_NAME}: {url}: {msg}"
             )
 
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return {"message": return_msg}, 500
 
         if item is None:
             msg = f"{self.model_name} with {kdict} not found"
-            logger.debug(msg)
+            current_app.logger.debug(msg)
             return {"message": msg}, 404
 
         adjust_before = self.before_commit.get(FUNC_NAME)
@@ -572,7 +582,7 @@ class ModelResource(DBBaseResource):
         except Exception as err:
             self.model_class.db.session.rollback()
             msg = err.args[0]
-            logger.error(f"{url} method {FUNC_NAME}: {msg}")
+            current_app.logger.error(f"{url} method {FUNC_NAME}: {msg}")
             return (
                 {
                     "message": "An error occurred deleting the "