+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## (0.3.6) -
+### Changed
+*   Added Python version specifications through Python 3.9.
+*   Changed logging to stem from Flask.current_app.
+
 ## (0.3.5) -
 ### Changed
 *   Specified a version restriction for SQLAlchemy to be lower than 1.4. Converted config :memory: to sqlite:///:memory:.